Optimize Megan Fox

+++ catscripts/Entertainment/Actors_and_actresses/Megan_Fox/MeganFox.sh
@@ -1,16 +1,20 @@
 #!/bin/bash
 
-export  KEYWORDS_ACTRESS_MEGANFOX="Megan(| )Fox"
+KEYWORDS_ACTRESS_MEGANFOX="Megan(| )Fox"
 
-egrep -i "$KEYWORDS_ACTRESS_MEGANFOX" newpages.txt >> MeganFox.txt
-
-MEGANFOX=`stat --print=%s MeganFox.txt`
-
-if [ $MEGANFOX -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="MeganFox.txt"
-  export CATNAME="Megan Fox"
-  $CATEGORIZE
-fi
 
-rm MeganFox.txt
+  MEGANFOX=`egrep -i "$KEYWORDS_ACTRESS_MEGANFOX" newpages.txt`
+
+  if [ "$MEGANFOX" != "" ];
+  then
+    egrep -i "$KEYWORDS_ACTRESS_MEGANFOX" newpages.txt > MeganFox.txt
+    export CATFILE="MeganFox.txt"
+    export CATNAME="Megan Fox"
+    $CATEGORIZE
+    rm MeganFox.txt
+    unset MEGANFOX
+  fi
+
+fi
